move tests into block mod and fix up use
